Refactor
Refactor folder and file structure
- Remove unused files
