+++ ui/lib/src/models/settings_model.dart
