--- ui/lib/src/pages/settings/constants.dart
@@ -1,10 +1,0 @@
-/// Class is for use with settings
-/// Could use instead for some part of user login?
-class Constants {
-  static const String settings = 'Settings';
-  static const String subscribe = "Subscribe";
-  static const String signOut = 'Sign Out';
-
-  // Create list of resources
-  static const List<String> choices = <String>[settings, subscribe, signOut];
-}
